requirements.txt
Flask
flask
requests
python-dotenv
flask-login
Flask-WTF
WTForms
bcrypt


